ml_services/requirements.txt
# ml_services/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
sentence-transformers==2.7.0
scikit-learn==1.3.0
psycopg2-binary==2.9.9
torch==2.3.1
numpy==1.26.4
python-dotenv